lstadd_files

+++ ftls_list.c
@@ -6,7 +6,7 @@
 /*   By: hvillasa <hvillasa@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/04/18 18:04:44 by hvillasa          #+#    #+#             */
-/*   Updated: 2017/04/24 08:13:37 by n                ###   ########.fr       */
+/*   Updated: 2017/04/24 08:34:24 by n                ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -52,9 +52,9 @@ void lstadd_files(t_files **head, char *file_name, t_flags flags)
 	t_new = (t_files*)malloc(sizeof(t_files));
 
 	t_new->file = ft_strdup(file_name);
-	t_new->st_mode = sb.st_mode;
-	t_new->st_size = sb.st_size;
-	
+	t_new->st_mode = (short)sb.st_mode;
+	t_new->st_size = (long)sb.st_size;
+	///ctime(sb.st_mtime, t_new->st_mtime);
 	if (LONG_FORMAT & flags)
 	{
 		ft_strcpy(t_new->fileprotection, fileprotection(sb.st_mode));
